// website/lib/services/search/search.service.type.ts
export interface SearchQuery {
  searchText: string;
}

export interface SearchOverviewResult {
  serchOverView: [];
  path: string;
}

export interface SearchObjectiveResult {
  serchRabRoleOrAdviceItem: [];
  path: string;
}

export interface SearchMobileAppFeatureResult {
  searchMobileAppFeature: [];
  path: string;
}

export interface SearchMemberResult {
  serchMember: [];
  path: string;
}

export interface SearchNocResult {
  serchNoc: [];
  path: string;
}